+++ src/tp2/CurvarEsquerda.java
@@ -3,10 +3,10 @@ package tp2;
 public class CurvarEsquerda extends Mensagem {
 
 	int raio;
-	int angulo;
+	float angulo;
 	
-	public CurvarEsquerda(int raio, int angulo) {
-		super(TIPO_MENSAGEM.CURVA_ESQ, String.format("Curva à Esquerda (raio %d, ângulo %d) \n", raio, angulo));
+	public CurvarEsquerda(int raio, float angulo) {
+		super(TIPO_MENSAGEM.CURVA_ESQ, String.format("Curva à Esquerda (raio %d, ângulo %d) \n", raio, (int)angulo));
 		this.raio = raio;
 		this.angulo = angulo;
 	}
@@ -15,7 +15,7 @@ public class CurvarEsquerda extends Mensagem {
 		return raio;
 	}
 
-	public int getAngulo() {
+	public float getAngulo() {
 		return angulo;
 	}
 }
